+++ omicverse/externel/flowsig/plotting/__init__.py
@@ -1,1 +1,1 @@
-from ._plotting import plot_intercellular_flows_from_inflows, plot_intercellular_flows_from_gems, plot_intercellular_flows_from_outflows
+from ._plotting import subset_for_flow_type,plot_intercellular_flows_from_inflows, plot_intercellular_flows_from_gems, plot_intercellular_flows_from_outflows